Moved the function to a header file and cleaned up some stuff
